% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{check_url}
\alias{check_url}
\title{Check if a URL is online}
\usage{
check_url(url, timeout_ms = 5000)
}
\arguments{
\item{url}{link to check}

\item{timeout_ms}{timeout in milliseconds}
}
\value{
list with online status, status code, and message
}
\description{
Check if a URL is online by making a HEAD request. Returns TRUE if the URL is
reachable and the status code is between 200 and 399 (inclusive). Returns FALSE
otherwise.
}
\details{
This function is useful for checking the availability of a server or
resource before attempting to access it. In particular, it is used by
\code{ed_info()} to check if the ERDDAP server is online before making requests
to it since the underlying \code{rerddap::info()} crashes R if \code{url} is offline.
}
\examples{
check_url("https://coastwatch.pfeg.noaa.gov/erddap/griddap/NOAA_DHW.html")
}
\concept{read}
