% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ed_dim}
\alias{ed_dim}
\title{Get dimension values from ERDDAP dataset}
\usage{
ed_dim(ed, dim)
}
\arguments{
\item{dim}{dimension to extract}

\item{ed_info}{ERDDAP info object on SeaScape dataset, as returned by \code{\link{ed_info}})}
}
\value{
vector of values for given dimension
}
\description{
Given an ERDDAP dataset info object, return a vector of all available values in given dimension.
}
\examples{
ed <- ed_info("https://coastwatch.pfeg.noaa.gov/erddap/griddap/jplMURSST41.html")
ed_dim(ed, "LEV")
}
\concept{read}
