% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ed_dims}
\alias{ed_dims}
\title{Get dimensions from ERDDAP dataset}
\usage{
ed_dims(ed)
}
\arguments{
\item{ed}{ERDDAP info object, as returned by \code{\link{ed_info}})}
}
\value{
character vector of dimension names
}
\description{
Given an ERDDAP dataset info object, return a character vector of all available dimensions.
}
\examples{
ed <- ed_info("https://coastwatch.pfeg.noaa.gov/erddap/griddap/jplMURSST41.html")
ed_dims(ed)
}
\concept{read}
