% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ed_info}
\alias{ed_info}
\title{Get ERDDAP dataset information}
\usage{
ed_info(dataset)
}
\arguments{
\item{dataset}{\verb{\{region\}_\{frequency\}} of dataset. Valid values (so far): "global_8day" or "global_monthly" (default).}
}
\value{
ERDDAP \code{\link[rerddap]{info}} object
}
\description{
Get ERDDAP dataset information.
}
\examples{
ed_info() # default: dataset = "global_monthly"
ed_info("global_8day")
}
\concept{read}
