% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ed_vars}
\alias{ed_vars}
\title{Get variables from ERDDAP dataset}
\usage{
ed_vars(ed)
}
\arguments{
\item{ed}{ERDDAP info object, as returned by \code{\link{ed_info}})}
}
\value{
character vector of variable names
}
\description{
Given an ERDDAP dataset info object, return a character vector of all available
variables. Variables are values that vary along the dimensions of the ERDDAP
datasets (as returned by \code{ed_dims()}).
}
\examples{
ed <- ed_info("https://coastwatch.pfeg.noaa.gov/erddap/griddap/jplMURSST41.html")
ed_vars(ed)
}
\concept{read}
