% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract text from a single pdf document}
\usage{
extract(path, which = "xpdf", ...)
}
\arguments{
\item{path}{(character) Path to a file}

\item{which}{(character) One of rcamp, gs, xpdf (default), or poppler}

\item{...}{further args passed on}
}
\value{
An object of class \code{rcamp_extr}, \code{gs_extr}, \code{xpdf_extr},
or \code{poppler_extr}. All share the same global class \code{extr}
}
\description{
This function wraps many methods to extract text from
non-scanned PDFs - no OCR methods used here. Available methods include
xpdf, Ghostscript, the Rcampdf package, and Poppler.
}
\examples{
\donttest{
path <- system.file("examples", "example1.pdf", package = "extractr")

# xpdf
xpdf <- extract(path, "xpdf")
xpdf$meta
xpdf$data

# Ghostscript
gs <- extract(path, "gs")
gs$meta
gs$data

# Rcampdf
rcamp <- extract(path, "rcamp")
rcamp$meta
rcamp$data

# Poppler
poppler <- extract(path, "poppler")
poppler$meta
poppler$data
}
}

