% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractr-package.R
\docType{package}
\name{extractr-package}
\alias{extractr}
\alias{extractr-package}
\title{R client for pdf extraction}
\description{
R interface to many web and local pdf extraction tools.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

