% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractr-package.R
\name{extractr_deps}
\alias{extractr_deps}
\title{extractr dependencies}
\description{
You can optionally use \code{Rcampdf} and \code{Rpoppler} in 
this package.  These two packages are not available easily via 
\code{\link{install.packages}}, so we put them in \code{Suggests} in this 
package, so are only required if they are chosen as an option in 
function calls.
}
\section{Rcampdf}{

You can get get \code{Rcampdf} by doing 
\code{install.packages("Rcampdf", repos = "http://datacube.wu.ac.at/", 
type = "source")}
}

