% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_ed_dates}
\alias{get_ed_dates}
\title{Get date range of Seascape dataset}
\usage{
get_ed_dates(ed_info)
}
\arguments{
\item{ed_info}{ERDDAP info object, as returned by \code{\link{get_ed_info}})}
}
\value{
date range of min and max
}
\description{
Get date range of Seascape dataset
}
\examples{
ed_info <- get_ed_info()
get_ed_dates(ed_info)
}
\concept{read}
