% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_ed_dates_all}
\alias{get_ed_dates_all}
\title{Get list of all dates available from Seascape dataset}
\usage{
get_ed_dates_all(ed_info, date_beg, date_end)
}
\arguments{
\item{ed_info}{ERDDAP info object on SeaScape dataset, as returned by \code{\link{get_ed_info}})}

\item{date_beg}{begin date}

\item{date_end}{end date}
}
\value{
vector of dates available
}
\description{
Given a SeaScape dataset info object and date range, return a vector of all available dates.
}
\examples{
ed_i <- ed_info()
get_ed_dates_all(ed_i, "2003-01-01", "2005-01-01")
}
\concept{read}
