% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_ed_grds}
\alias{get_ed_grds}
\title{Get Seascape grids within polygon for date range}
\usage{
get_ed_grds(
  ed_info,
  ply,
  ed_var = "CLASS",
  date_beg = min(get_ed_dates(ed_info)),
  date_end = max(get_ed_dates(ed_info)),
  dir_tif = NULL,
  del_cache = F,
  verbose = T
)
}
\arguments{
\item{ed_info}{SeaScape ERDDAP info object, as returned by
\code{\link{ed_info}})}

\item{ply}{polygon as spatial feature \code{\link[sf]{sf}}, as returned by
\code{\link{get_url_ply}} or \code{\link{bbox_to_ply}}}

\item{ed_var}{SeaScape variable. One of "CLASS" (default) or "P" for
probability.}

\item{date_beg}{date begin to fetch, as character (\code{"2003-01-15"}) or Date
(\code{Date("2003-01-15")}). Defaults to first date available from \code{ed_info}.}

\item{date_end}{date end to fetch, as character (\code{"2020-11-15"}) or Date
(\code{Date("2020-11-15")}). Defaults to latest date available from \code{ed_info}.}

\item{dir_tif}{directory to cache results. Files are stored in the format
\verb{grd_\{ed_var\}_\{date\}.tif} so any other information needed, such as place or
SeaScape dataset should be captured by the containing folders so as to not
inadvertently write or read the wrong grid. This folder is consulted for
available dates before fetching any missing from the ERDDAP server.}

\item{del_cache}{Delete ERDDAP cache with
\code{\link[rerddap]{cache_delete_all}}, which may be necessary if
underlying data changed. Default: FALSE.}

\item{verbose}{display messages on status of function. Useful for debugging
or showing status while getting data from a wide range and/or big polygon.
Default: FALSE.}
}
\value{
Raster \code{\link[raster]{raster}} layer if one date,
\code{\link[raster]{stack}} if more
}
\description{
Given a polygon and date range, fetch Seascape data and return a raster layer
for a single date or raster stack if multiple dates found for given date
range.
}
\examples{
ply  <- get_url_ply("mbnms")
ed_i <- ed_info()
grds <- get_ed_grds(ed_i, ply, date_beg = "2020-01-01")
grds

}
\concept{read}
