% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_ed_info}
\alias{get_ed_info}
\title{Get Seascape dataset information}
\usage{
get_ed_info(dataset)
}
\arguments{
\item{dataset}{\verb{\{region\}_\{frequency\}} of dataset. Valid values (so far): "global_8day" or "global_monthly" (default).}
}
\value{
ERDDAP \code{\link[rerddap]{info}} object
}
\description{
Get Seascape dataset information from ERDDAP server.
}
\examples{
get_ed_info() # default: dataset = "global_monthly"
get_ed_info("global_8day")
}
\concept{read}
