% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_ed_vals_all}
\alias{get_ed_vals_all}
\title{Get list of all dates available from Seascape dataset}
\usage{
get_ed_vals_all(ed_info, var)
}
\arguments{
\item{ed_info}{ERDDAP info object on SeaScape dataset, as returned by \code{\link{get_ed_info}})}

\item{var}{variable to extract}
}
\value{
vector of values for given variable
}
\description{
Given a SeaScape dataset info object and date range, return a vector of all available dates.
}
\examples{
ed_i <- get_ed_info("https://apdrc.soest.hawaii.edu/erddap/griddap/hawaii_soest_d749_a206_cd3a.html")
get_ed_vals_all(ed_i, "LEV")
}
\concept{read}
