% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData.fn}
\alias{getData.fn}
\title{Create an ODBC connection and get data from database}
\usage{
getData.fn(dsn, uid, query, pw = "", ais = asis)
}
\arguments{
\item{dsn}{The name of your databse connection}

\item{uid}{Your user ID}

\item{query}{The sql query as an R object}

\item{pw}{The password for the database connection. If omitted,
RODBC will prompt you for a password and will mask it.
If entered, it will be visible and unsecure.}

\item{ais}{Logical whether or not to convert columns as in \code{\link{read.table}}.
             See \code{\link{RODBC::sqlQuery}}}
}
\description{
Create an ODBC connection and get data from database
}
\details{
Open an ODBC connection and using an sql, extract data from the database.
}
\author{
Andi Stephens

Allan Hicks
}
\seealso{
\code{\link{make.query}, \link{queryDB}}, \code{\link{getMultiSpeciesCatch}}, \code{\link{RODBC::sqlQuery}}
}

