% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TR_getMultiSpeciesCatch.R
\name{getMultiSpeciesCatch}
\alias{getMultiSpeciesCatch}
\title{Wrapper to make a query from a sql file and get the data from a database using RODBC
A TrawlR function for use with the FRAM trawl survey database.}
\usage{
getMultiSpeciesCatch(queryFN, db, uid, pw = "", sp, start, end,
  speciesTable = NULL, asis = F)
}
\arguments{
\item{queryFN}{The filename and location of the sql query to read in and convert with \code{\link{make.query}}}

\item{db}{The name of your databse connection}

\item{uid}{your user ID}

\item{pw}{The password for the database connection. If omitted,
RODBC will prompt you for a password and will mask it.
If entered, it will be visible and unsecure.}

\item{sp}{The species scientific name to extract from the DB}

\item{start}{The start year}

\item{end}{The end year}

\item{speciesTable}{The end year}

\item{ais}{Logical whether or not to convert columns as in \code{\link{read.table}}.
             See \code{\link{RODBC::sqlQuery}}}
}
\description{
Wrapper to make a query from a sql file and get the data from a database using RODBC
A TrawlR function for use with the FRAM trawl survey database.
}
\details{
Calls \code{\link{make.query}} and then \code{\link{getData}}
}
\author{
Andi Stephens

Allan Hicks
}
\seealso{
\code{\link{make.query}, \link{queryDB}}, \code{\link{getData}}, \code{\link{RODBC::sqlQuery}}
}

