% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSQL.R
\name{getSQL}
\alias{getSQL}
\title{Download sql file from GitHub and save as an object in R}
\usage{
getSQL(sqlName, folder = "GitHub", database = if (folder == "GitHub") {    
  c("trawl", "observer") } else {     "" }, verbose = TRUE)
}
\arguments{
\item{sqlName}{The name of your sql file. This can be just the name, or a specific path and name on your computer. See Details.}

\item{folder}{The folder your sql file is in. "GitHub" refers to the specific GitHub site (nwfsc-data)}

\item{database}{An additional folder specifying the database that the sql file is associated with. Can be a vector and it
                  will search through these folders until it finds the sql file (if present)}
}
\description{
Download sql file from GitHub and save as an object in R
}
\details{
Reads an sql (text) file from GitHub or any folder for use in ExtractR
}
\examples{
getSQL("HaulCharacteristics.sql") #gets the sql from GitHub by default

getSQL(sqlName = "HaulCharacteristics.sql",
	   folder = "C:/NOAA2015/SurveyDatabase/sql",
	   database = "")

getSQL("C:/NOAA2015/SurveyDatabase/sql/HaulCharacteristics.sql")
}
\author{
Allan Hicks
}
\seealso{
\code{\link{make.query}, \link{queryDB}}, \code{\link{getMultiSpeciesCatch}}, \code{\link{RODBC::sqlQuery}}
}

