% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{grds_to_ts}
\alias{grds_to_ts}
\title{Summarize grids with a continuous variable into a time series table}
\usage{
grds_to_ts(grds, fxns = c("mean", "sd"), ts_csv = NULL, verbose = F)
}
\arguments{
\item{grds}{raster stack with more than one date, as returned by
\code{\link{get_ss_grds}}}

\item{ts_csv}{path to csv to save this time series table. Default is NULL, in
which case the table is not saved. If path is set and already exists then
that will be read in if all dates in the \code{grds} are present instead of
recalculating and writing to \code{ts_csv}.}

\item{verbose}{show messages of process. Useful for debugging. Default:
FALSE.}
}
\value{
\code{tibble} of data
}
\description{
Summarize grids into a table having columns: \code{date}, \code{cellvalue} (e.g. Seascape
CLASS) and number of cells (\code{n_cells}).
}
\details{
This function is particularly helpful in between using \code{get_ed_grds()} and
\code{plot_ts_cat()}.
}
\examples{
ply  <- get_url_ply("mbnms")
ss_i <- get_ss_info()
grds <- get_ss_grds(ss_i, ply, date_beg = "2020-01-01")
tbl  <- sum_ss_grds_to_ts(grds)
tbl

}
\concept{analyze}
