% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make.query.R
\name{make.query}
\alias{make.query}
\title{Make a query from an sql file that can be used by RODBC}
\usage{
make.query(query, sp = " ", start = " ", end = " ", ...)
}
\arguments{
\item{query}{The sql file.}

\item{sp}{The species scientific name to extract from the DB (certain sql scripts)}

\item{start}{The start year (certain sql scripts)}

\item{end}{The end year (certain sql scripts)}

\item{...}{Named arguments to pass to \code{\link{getSQL}}. See examples.}
}
\description{
Make a query from an sql file that can be used by RODBC
}
\details{
Read in the lines of an sql file, remove the comments, substitute in stuff,
  and then return the query as a character string.
}
\examples{
make.query("HaulCharacteristics.sql", start=1998, end=2014) #gets the sql from GitHub by default

make.query("SingleSpecies.sql", sp="Sebastes entomelas", start=1998, end=2014) #gets the sql from GitHub by default

make.query("HaulCharacteristics.sql", start=1998, end=2014,
    folder = "C:/NOAA2015/SurveyDatabase/sql",
    database = "")

make.query("C:/NOAA2015/SurveyDatabase/sql/HaulCharacteristics.sql", start=1998, end=2014)
}
\author{
Andi Stephens

Allan Hicks
}
\seealso{
\code{\link{getData}, \link{queryDB}}, \code{\link{getMultiSpeciesCatch}}
}

