% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdf.r
\name{pdftotext}
\alias{pdftotext}
\title{Use pdftotext to get text from a pdf}
\usage{
pdftotext(path, ...)
}
\arguments{
\item{path}{A file path.}

\item{...}{Further command line flags passed to pdftotext. see examples}
}
\description{
Use pdftotext to get text from a pdf
}
\examples{
\donttest{
path <- "~/github/sac/scott/pdfs/ChamberlainEtal2013Ecosphere.pdf"
pdftotext(path)
}
}
