% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rcampdf}
\alias{rcampdf}
\title{PDF-to-text conversion using Rcampdf}
\usage{
rcampdf(files = NULL, ...)
}
\arguments{
\item{files}{(character) Path to a file, or files on your machine.}

\item{...}{Further args passed on to the \code{tm} package, which does the conversion}
}
\value{
A S3 object of class xpdf with slots for meta and data
}
\description{
Uses a local tool called Rcamppdf at \url{http://datacube.wu.ac.at/}.
}
\examples{
\dontrun{
path <- "~/github/sac/scott/pdfs/ChamberlainEtal2013Ecosphere.pdf"
rcampdf(files = path)

paths <- c("~/github/sac/scott/pdfs/BarraquandEtal2014peerj.pdf",
"~/github/sac/scott/pdfs/Chamberlain&Holland2009Ecology.pdf",
"~/github/sac/scott/pdfs/Revell&Chamberlain2014MEE.pdf")
(res <- rcampdf(files=paths))
res$data[[3]]
}
}
\author{
Scott Chamberlain {myrmecocystus@gmail.com}
}

