% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdf.r
\name{xpdf}
\alias{xpdf}
\title{PDF-to-text conversion using XPDF}
\usage{
xpdf(files = NULL, ...)
}
\arguments{
\item{files}{(character) Path to a file, or files on your machine.}

\item{...}{Further args passed on to the \code{tm} package, which does 
the conversion}
}
\value{
A S3 object of class xpdf with slots for meta and data
}
\description{
PDF-to-text conversion using XPDF
}
\examples{
\dontrun{
path <- "~/github/sac/scott/pdfs/ChamberlainEtal2013Ecosphere.pdf"
res <- xpdf(files = path)

paths <- c("~/github/sac/scott/pdfs/BarraquandEtal2014peerj.pdf", 
"~/github/sac/scott/pdfs/Chamberlain&Holland2009Ecology.pdf",
"~/github/sac/scott/pdfs/Revell&Chamberlain2014MEE.pdf")
(res <- xpdf(files=paths))
res$data[[3]]
}
}
\author{
Scott Chamberlain {myrmecocystus@gmail.com}
}
