% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ed_extract}
\alias{ed_extract}
\title{ERDDAP Extract}
\usage{
ed_extract(
  ed,
  var,
  sf_zones = NULL,
  fld_zones = NULL,
  bbox = NULL,
  zonal_csv,
  zonal_fun = "mean",
  rast_tif = NULL,
  mask_tif = TRUE,
  dir_nc = NULL,
  keep_nc = FALSE,
  n_max_vals_per_req = 1e+05,
  n_max_retries = 3,
  time_min = NULL,
  time_max = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{ed}{ERDDAP info object, of class \code{rerddap::info} as returned by \code{\link{ed_info}})}

\item{var}{variable to extract}

\item{sf_zones}{spatial feature object (\code{\link{sf}}) with zones to extract zonal
statistics from ERDDAP gridded data. If \code{sf_zones} is left
to the default \code{NULL} value, then the \code{bbox} is used.}

\item{fld_zones}{character vector of unique field name(s) in \code{sf_zones} to
include in extracted zonal statistics}

\item{bbox}{bounding box (e.g.,
\code{c(xmin = -83.0, ymin = 27.3, xmax = -81.8, ymax= 28.5)}), used to extract
the grid (i.e., as \code{longitude} and \code{latitude}  arguments to \code{rerddap::griddap()}).
If \code{bbox} is left to the default \code{NULL} value, then the bounding box is derived from the bounding box of the
\code{sf_zones}.}

\item{zonal_csv}{output zonal statistics as CSV from \code{terra::zonal(rast_tif, aoi)}}

\item{zonal_fun}{function to summarize the data over time by \code{sf_zones}. Default:}

\item{rast_tif}{optional output as GeoTIFF masked to \code{sf_zones} from NetCDF}

\item{mask_tif}{mask the GeoTIFF by \code{sf_zones}. Default: TRUE}

\item{dir_nc}{optional output directory to keep NetCDF files written to disk
by \code{rerddap::griddap()}}

\item{keep_nc}{keep NetCDF files written to disk by \code{rerddap::griddap()}
server at a time. Default: 100,000}

\item{n_max_vals_per_req}{maximum number of values to request from ERDDAP}

\item{n_max_retries}{maximum number of retries to request data from ERDDAP
server}

\item{time_min}{minimum time to extract from ERDDAP dataset}

\item{time_max}{maximum time to extract from ERDDAP dataset}

\item{verbose}{display messages on status of function. Useful for debugging
or showing status while getting data from a wide range and/or big polygon.
Default: FALSE.}

\item{...}{arguments to pass along to \code{rerddap::griddap()}, such as to
filter the request by dimensions}
}
\value{
invisible first argument (\code{ed} object), since called for side effects
}
\description{
Extract data from ERDDAP dataset by bounding box and/or polygon. This function
provides the following enhancements to \code{rerddap::griddap()}:
\enumerate{
\item Throttle request to \code{rerddap::griddap()} by chunks of time, so
the ERDDAP server does not become unresponsive with too big a request while
minimizing the number of requests.
\item Mask the grids by Area of Interest (\code{aoi}) (using \code{terra::mask()}).
\item Summarize grids over time by \code{aoi}, including sub-geometries, into a CSV (using \code{terra::zonal()}).
\item Preserve intermediary NetCDF and/or GeoTIFF files for raster visualization.
}
}
\examples{
ed <- ed_info("https://coastwatch.pfeg.noaa.gov/erddap/griddap/jplMURSST41.html")
(vars <- ed_vars(ed))
ed_extract(
  ed,
  "analysed_sst",
  bbox = c(xmin = -83.0, ymin = 27.3, xmax = -81.8, ymax= 28.5))
}
