% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{extract-package}
\alias{extract}
\alias{extract-package}
\title{R client for pdf extraction}
\description{
R interface to many web and local pdf extraction tools.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

