% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract text from a single pdf document}
\usage{
extract(paths, which = "xpdf", ...)
}
\arguments{
\item{paths}{(character) One or more paths to a file}

\item{which}{(character) One of rcamp, gs, xpdf (default), or poppler}

\item{...}{further args passed on}
}
\value{
A list or a single object, of class \code{rcamp_extr}, \code{gs_extr},
\code{xpdf_extr}, or \code{poppler_extr}. All share the same global class \code{extr}
}
\description{
This function wraps many methods to extract text from
non-scanned PDFs - no OCR methods used here. Available methods include
xpdf, Ghostscript, the Rcampdf package, and Poppler.
}
\examples{
\dontrun{
path <- system.file("examples", "example1.pdf", package = "extractr")

# xpdf
xpdf <- extract(path, "xpdf")
xpdf$meta
xpdf$data

# Ghostscript
gs <- extract(path, "gs")
gs$meta
gs$data

# Rcampdf
rcamp <- extract(path, "rcamp")
rcamp$meta
rcamp$data

# Poppler
poppler <- extract(path, "poppler")
poppler$meta
poppler$data

# Pass many paths at once
path1 <- system.file("examples", "example1.pdf", package = "extractr")
path2 <- system.file("examples", "example2.pdf", package = "extractr")
path3 <- system.file("examples", "example3.pdf", package = "extractr")
extract(c(path1, path2, path3))
}
}

