% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractr-package.R
\docType{package}
\name{extractr-package}
\alias{extractr}
\alias{extractr-package}
\title{R client for pdf extraction}
\description{
R interface to many web and local pdf extraction tools.
}
\section{Dependencies}{


See \code{\link{extractr_deps}} for information on dependencies. Most packages and
tools are easy, but some require some more in depth instructions.
}

\section{Tools used in extractr}{


\itemize{
 \item XPDF - \url{http://www.foolabs.com/xpdf/} - Local usage
 \item Ghostscript - \url{http://www.ghostscript.com} - Local usage
 \item Rcampdf - \url{http://datacube.wu.ac.at/} - Local usage
 \item Poppler - \url{http://poppler.freedesktop.org/} - Local usage
 \item pdfx - \url{http://pdfx.cs.man.ac.uk/usage} - Web API
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

