% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.query.R
\name{make.query}
\alias{make.query}
\title{Make a query from an sql file that can be used by RODBC}
\usage{
make.query(query, sp = " ", start = " ", end = " ")
}
\arguments{
\item{query}{The sql file.}

\item{sp}{The species scientific name to extract from the DB}

\item{start}{The start year}

\item{end}{The end year}
}
\description{
Make a query from an sql file that can be used by RODBC
}
\details{
Read in the lines of an sql file, remove the comments, substitute in stuff,
  and then return the query as a character string.
}
\author{
Andi Stephens

Allan Hicks
}
\seealso{
\code{\link{getData}, \link{queryDB}}, \code{\link{getMultiSpeciesCatch}}
}

