% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series}
\usage{
plot_ts(
  ts_csv,
  fld_avg = "mean",
  fld_sd = NULL,
  fld_date = "date",
  color = "red",
  label = "Temperature (°C)",
  ...
)
}
\arguments{
\item{ts_csv}{file path to timeseries table as comma-separated value (CSV) (required)}

\item{fld_avg}{field name containing value average; default = \code{"mean"}}

\item{fld_sd}{field name containing standard deviation average (optional); default = \code{NULL}}

\item{fld_date}{field name containing date (required); default = \code{"date"}}

\item{color}{color to plot value; default = \code{"red"}}

\item{label}{label for y-axes of plot}

\item{...}{additional parameters to pass to \code{dygraphs::dygraph()}}
}
\value{
\code{dygraphs::dygraph()} object of the time series plot
}
\description{
The purpose of this function is to generate time series plots of ERRDAP data.
}
\examples{
\dontrun{
ts_csv <- here::here("data_tmp/ts.csv")
plot_ts(ts_csv, main = "SST")
}

}
\concept{visualize}
